pyproject.toml
[project]
name = "dspy-data"
version = "0.1.0"
description = "Automated data collection using dspy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy-ai>=3.0.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "ruff>=0.12.9",
]
